feat: add generate key button in add key popup
