+++ extension/chrome/settings/index.htm
@@ -203,13 +203,6 @@
                   <a href="#" class="show_settings_page my_key_link add_key" data-test="action-open-add-key-page" page="/chrome/settings/modules/add_key.htm"
                     >Add Key</a
                   >
-                  <a
-                    href="#"
-                    class="show_settings_page my_key_link generate_key"
-                    data-test="action-open-generate-key-page"
-                    page="/chrome/settings/modules/generate_key.htm"
-                    >Generate Key</a
-                  >
                 </span>
               </div>
             </div>
